// src/infrastructure/types/types.ts
export type UserRegDTO = {
    login: string
    password: string
    email: string
}
export type AuthDTO = {
    loginOrEmail: string
    password: string
}
export type PassCodeDTO = {
    recoveryCode: string
    newPassword: string
}

// ResBody - типизация ответа
// Locals - возможно переменные окружения
// e.Request<P,ResBody,ReqBody,ReqQuery,Locals>
//     <Params, ??????,   Body,   Query, ?????>

export type UserIdModel = {
    userId?: string |  null
}

export type ReqParamsType<T>         = Request<T>
export type ReqBodyType<T>           = Request<{},{},T>